--- adj.txt
@@ -1,2 +1,0 @@
-a
-1
