+++ insults.json
@@ -1,0 +1,8 @@
+{
+    "adj": [" stupid", " stinky", " useless", "n Aloof", "n Arrogant", " Belligerent", " Big-headed", " Boastful", " Bossy", " Callous", " Careless", " Clingy", " Confrontational", " Cowardly", " Cruel", " Cynical", " Defensive", " Finicky", " Foolish", " Grumpy", " Gullible", " Hostile", "n Idle", "n Impolite", "n Inconsiderate", "n Indecisive", "n Irresponsible", " Moody", " Nasty", " Overcritical", " Overemotional", " Pessimistic", " Pompous", " Self-centered", " Silly", " Sullen", " Thoughtless", " Touchy", "n Untidy", " Vain", " Weak-willed"],
+
+    "verbs": ["Fucking", "Shitting", "Pissing", "Bitching", "Cock-sucking", "Ass-kissing", "Jacking off", "Masturbating", "Rimming", "Choking", "Whipping", "Spanking", "Slapping", "Teasing", "Taunting", "Mocking", "Insulting", "Belittling", "Degrading", "Slandering", "Defaming", "Trashing", "Bashing", "Humiliating", "Ridiculing", "Mocking", "Sneering", "Smirking", "Snickering", "Jeering", "Scoffing", "Disparaging", "Disrespecting", "Disparaging", "Devaluing", "Deprecating", "Denigrating", "Debasing", "Condemning", "Vilifying", "Cursing", "Swearing", "Name-calling", "Taunting", "Provoking", "Aggravating", "Harassing", "Bullying", "Threatening", "Intimidating"],
+
+    "nouns": ["twat", "dumbass", "furry", "Asshole", "Bastard", "Bitch", "Cock", "Cunt", "Dick", "Douchebag", "Faggot", "Fuckface", "Jerk", "Motherfucker", "Pussy", "Shithead", "Slag", "Slut", "Twat", "Wanker", "Arsehole", "Bollocks", "Clusterfuck", "Cocksucker", "Crapbag", "Dipshit", "Douchecanoe", "Dumbass", "Fucker", "Gigolo", "Gooch", "Jackass", "Numbnuts", "Pissface", "Prick", "Ratfink", "Shitbag", "Skank", "Sleazebag", "Snotrocket", "Son of a bitch", "Turd", "Asswipe", "Butt munch", "Camel jockey", "Chicken fucker", "Chode", "Clusterfucker", "Cumstain", "Dingleberry", "Flaming asshole", "Shitstain", "Taint", "snowflake"]
+
+}
