+++ main.py
@@ -1,0 +1,3 @@
+filename = input()
+file2 = open(filename, "r").read().splitlines()
+print(file2)
