--- nou.txt
@@ -1,2 +1,0 @@
-b
-2
