+++ public/index.html
@@ -51,12 +51,7 @@
 
     }
 
-    function insultOnLoad() {
-        // Adding functionality from py.insultgen
-        socket.emit('insult', (callback) => {
-            console.log(callback)
-        })
-    }
+    
 </script>
 
 <style>
