--- ver.txt
@@ -1,2 +1,0 @@
-c
-3
